Handle purely real input.
Since numpy.real() and numpy.imag() also operate on floats and
numpy-arrays of floats, these types can be safely included in the
if-statement. When this is done, purely real input data (single floats
and numpy float arrays) is handled correctly.

Note that the code does not handle python lists of complex elements.
Thus, no support for python lists of float elements is added here,
although this could be done in the future. For now, python lists must
be converted to numpy-arrays.

This commit fixes Issue #3.